+++ phytreon/plot/__init__.py
@@ -9,6 +9,7 @@ from .tangle import TangleFigure
 from .densi import DensiTreeFigure
 from .network import SequenceNetwork
 from .panels import PanelFigure, panels
+from .elements import looks_like_a_taxon
 from .splitnet import (SplitNetwork, neighbor_net,
                        neighbornet_ordering)
 
@@ -20,6 +21,7 @@ __all__ = [
     "PanelFigure",
     "panels",
     "SplitNetwork", "neighbor_net", "neighbornet_ordering",
+    "looks_like_a_taxon",
     "RenderContext",
     "ColorScale",
     "build_color_scale",